# openapi.yaml
openapi: 3.0.3
info:
  title: Landed Cost Calculation API
  description: |
    API สำหรับคำนวณต้นทุนนำเข้าสินค้า (Landed Cost) จากจีนเข้าสหรัฐอเมริกา
    รวมภาษีศุลกากรหลายชั้น (MFN, Section 301, Section 232, IEEPA)
    พร้อมระบบ temporal versioning, exclusion tracking และ audit trail

    ## ภาษีที่รองรับ
    | ประเภท | คำอธิบาย | อัตราทั่วไป |
    |--------|----------|------------|
    | MFN | อัตราฐาน Most Favored Nation | 0-9% |
    | Section 301 | ภาษีเฉพาะจีน (2018+) | 25% |
    | Section 232 | ภาษีความมั่นคงแห่งชาติ (เหล็ก/อลูมิเนียม) | 25% |
    | IEEPA | ภาษีตอบโต้ reciprocal (2025+) | 14.5% |

    ## Stacking Logic
    - MFN เป็นฐานเสมอ
    - Section 301 และ IEEPA **เลือกอันที่สูงกว่าอันเดียว**
    - Section 232 และ IEEPA **เลือกอันที่สูงกว่าอันเดียว**
  version: 1.0.0
  contact:
    name: API Support
    email: support@landed-cost.dev
  license:
    name: MIT
servers:
  - url: https://landed-cost.fly.dev
    description: Production
  - url: https://landed-cost-staging.fly.dev
    description: Staging
  - url: http://localhost:8000
    description: Local Development
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /tariff-rules:
    get:
      tags: [Tariff Rules]
      summary: ดูรายการกฎภาษีทั้งหมด
      description: |
        แสดงกฎภาษีทั้งหมดในระบบ สามารถ filter ตามประเภทภาษีหรือรหัส HTS
        แต่ละกฎมี effective_from/effective_to สำหรับ temporal versioning
      operationId: listTariffRules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: tariff_type
          in: query
          description: กรองตามประเภทภาษี
          required: false
          schema:
            type: string
            enum: [MFN, SECTION_301, SECTION_232, IEEPA]
        - name: hts_code
          in: query
          description: กรองตาม HTS code pattern (เช่น '8483')
          required: false
          schema:
            type: string
            example: "8483"
        - name: sort
          in: query
          description: เรียงลำดับ
          required: false
          schema:
            type: string
            enum: [effective_from_asc, effective_from_desc, rate_asc, rate_desc]
            default: effective_from_desc
      responses:
        '200':
          description: รายการกฎภาษี
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TariffRuleResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: "301-8483"
                    tariff_type: SECTION_301
                    hts_code_pattern: "8483"
                    origin_country: CN
                    rate: 0.25
                    effective_from: "2018-07-06"
                    effective_to: null
                    stacks_with: [MFN]
                    mutually_exclusive_with: [IEEPA]
                    description: "Section 301 List 1 - 25% tariff on Chinese gears/shafts"
                    source_reference: "USTR FR Notice 83 FR 28710"
                meta:
                  page: 1
                  page_size: 20
                  total_items: 1
                  total_pages: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Tariff Rules]
      summary: สร้างกฎภาษีใหม่
      description: |
        เพิ่มกฎภาษีใหม่เข้าระบบ กฎจะมีผลเมื่อ:
        - HTS code ของสินค้าตรงกับ pattern (prefix matching)
        - ประเทศต้นทางตรงกัน
        - วันที่นำเข้าอยู่ในช่วง effective_from ถึง effective_to

        **การตั้ง stacking:**
        - `stacks_with`: ประเภทภาษีที่กฎนี้ซ้อนทับได้ (เช่น ["MFN"])
        - `mutually_exclusive_with`: ประเภทที่เลือกอันสูงกว่าอันเดียว (เช่น ["IEEPA"])
      operationId: createTariffRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TariffRuleCreate'
            example:
              tariff_type: SECTION_301
              hts_code_pattern: "8483"
              origin_country: CN
              rate: 0.25
              effective_from: "2018-07-06"
              effective_to: null
              stacks_with: [MFN]
              mutually_exclusive_with: [IEEPA]
              description: "Section 301 List 1 - 25% on Chinese machinery parts"
              source_reference: "USTR FR Notice 83 FR 28710"
      responses:
        '201':
          description: สร้างกฎภาษีสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TariffRuleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tariff-rules/{rule_id}:
    put:
      tags: [Tariff Rules]
      summary: แก้ไขกฎภาษีที่มีอยู่
      description: |
        อัปเดตกฎภาษี เฉพาะ field ที่ส่งมาเท่านั้นจะถูกแก้ไข
        ใช้สำหรับ: ปรับอัตรา, ตั้งวันหมดอายุ, เปลี่ยน stacking config
      operationId: updateTariffRule
      parameters:
        - name: rule_id
          in: path
          required: true
          description: UUID ของกฎภาษี
          schema:
            type: string
            format: uuid
            example: "301-8483"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TariffRuleUpdate'
            example:
              rate: 0.30
              description: "Updated: Section 301 increased to 30%"
      responses:
        '200':
          description: อัปเดตสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TariffRuleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /exclusions:
    get:
      tags: [Exclusions]
      summary: ดูรายการข้อยกเว้นภาษี
      description: |
        แสดง exclusion ทั้งหมด (ทั้งที่ยังมีผลและหมดอายุแล้ว)
        Exclusion คือการยกเว้นภาษีชั่วคราวสำหรับ HTS code เฉพาะ
        เช่น Semiconductor ICs ได้รับยกเว้น IEEPA ถาวร
      operationId: listExclusions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: hts_code
          in: query
          description: กรองตาม HTS code
          required: false
          schema:
            type: string
            example: "8542.31"
      responses:
        '200':
          description: รายการ exclusions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExclusionResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: "excl-8542-31"
                    hts_code: "8542.31"
                    tariff_type: IEEPA
                    origin_country: CN
                    effective_from: "2025-04-09"
                    effective_to: null
                    description: "Semiconductor ICs excluded from IEEPA reciprocal tariff"
                    source_reference: "Executive Order 14257 Annex II"
                meta:
                  page: 1
                  page_size: 20
                  total_items: 1
                  total_pages: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Exclusions]
      summary: สร้างข้อยกเว้นภาษีใหม่
      description: |
        เพิ่ม exclusion ใหม่ เมื่อ active (วันนำเข้าอยู่ในช่วง effective)
        ภาษีประเภทที่ระบุจะไม่ถูก apply กับ HTS code นั้น
        ตั้ง effective_to เป็น null สำหรับ exclusion ถาวร
      operationId: createExclusion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExclusionCreate'
            example:
              hts_code: "8471.30"
              tariff_type: SECTION_301
              origin_country: CN
              effective_from: "2025-04-11"
              effective_to: "2025-07-09"
              description: "Temporary exclusion for laptops from Section 301"
              source_reference: "USTR 90 FR 15658"
      responses:
        '201':
          description: สร้าง exclusion สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ExclusionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key สำหรับเข้าถึงระบบ ขอได้จากผู้ดูแลระบบ
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token สำหรับ production (future)
  parameters:
    PageParam:
      name: page
      in: query
      description: หมายเลขหน้า (เริ่มจาก 1)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PageSizeParam:
      name: page_size
      in: query
      description: จำนวนรายการต่อหน้า
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
  responses:
    BadRequest:
      description: ข้อมูลที่ส่งมาไม่ถูกต้อง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: BAD_REQUEST
              message: "HTS code '9999.99' not found in the system"
              details: null
    Unauthorized:
      description: ไม่มี API Key หรือ API Key ไม่ถูกต้อง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: "Missing API key. Provide it via the X-API-Key header."
              details: null
...
    NotFound:
      description: ไม่พบ resource ที่ร้องขอ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: NOT_FOUND
              message: "Resource not found"
              details: null
    ValidationError:
      description: ข้อมูลไม่ผ่าน validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: VALIDATION_ERROR
              message: "Request validation failed"
              details:
                - field: rate
                  message: "ensure this value is greater than or equal to 0"
    InternalServerError:
      description: เกิดข้อผิดพลาดภายในระบบ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: INTERNAL_ERROR
              message: "An unexpected error occurred"
              details: null
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: รหัส error มาตรฐาน
              example: BAD_REQUEST
            message:
              type: string
              description: คำอธิบาย error สำหรับ developer
              example: "HTS code '9999.99' not found"
            details:
              description: รายละเอียดเพิ่มเติม (validation errors, etc.)
              nullable: true
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 3
...
    TariffRuleCreate:
      type: object
      required: [tariff_type, hts_code_pattern, rate, effective_from]
      properties:
        tariff_type:
          type: string
          enum: [MFN, SECTION_301, SECTION_232, IEEPA]
          description: ประเภทภาษี
        hts_code_pattern:
          type: string
          description: HTS code หรือ prefix (เช่น '84' ครอบคลุมทั้ง chapter)
          minLength: 2
          maxLength: 20
          example: "8483"
        origin_country:
          type: string
          description: รหัสประเทศต้นทาง (ISO 3166-1 alpha-2)
          pattern: "^[A-Z]{2}$"
          default: CN
          example: CN
        rate:
          type: number
          format: double
          description: อัตราภาษี (0.25 = 25%)
          minimum: 0
          maximum: 5.0
          example: 0.25
        effective_from:
          type: string
          format: date
          description: วันที่เริ่มมีผล
          example: "2018-07-06"
        effective_to:
          type: string
          format: date
          nullable: true
          description: วันที่หมดอายุ (null = ยังมีผลอยู่)
          example: null
        stacks_with:
          type: array
          items:
            type: string
          description: ประเภทภาษีที่กฎนี้ซ้อนทับได้
          example: [MFN]
        mutually_exclusive_with:
          type: array
          items:
            type: string
          description: ประเภทที่เลือกอันสูงกว่าอันเดียว
          example: [IEEPA]
        description:
          type: string
          maxLength: 500
          example: "Section 301 List 1 - 25% on Chinese machinery"
        source_reference:
          type: string
          maxLength: 500
          description: อ้างอิงกฎหมาย/ประกาศ
          example: "USTR FR Notice 83 FR 28710"
    TariffRuleUpdate:
      type: object
      description: เฉพาะ field ที่ส่งมาจะถูกอัปเดต
      properties:
        rate:
          type: number
          format: double
          minimum: 0
          maximum: 5.0
          nullable: true
        effective_to:
          type: string
          format: date
          nullable: true
        stacks_with:
          type: array
          items:
            type: string
          nullable: true
        mutually_exclusive_with:
          type: array
          items:
            type: string
          nullable: true
        description:
          type: string
          maxLength: 500
          nullable: true
        source_reference:
          type: string
          maxLength: 500
          nullable: true
    TariffRuleResponse:
      type: object
      properties:
        id:
          type: string
        tariff_type:
          type: string
          enum: [MFN, SECTION_301, SECTION_232, IEEPA]
        hts_code_pattern:
          type: string
        origin_country:
          type: string
        rate:
          type: number
          format: double
        effective_from:
          type: string
          format: date
        effective_to:
          type: string
          format: date
          nullable: true
        stacks_with:
          type: array
          items:
            type: string
        mutually_exclusive_with:
          type: array
          items:
            type: string
        description:
          type: string
        source_reference:
          type: string
    ExclusionCreate:
      type: object
      required: [hts_code, tariff_type, effective_from]
      properties:
        hts_code:
          type: string
          description: HTS code ที่ได้รับยกเว้น
          minLength: 2
          maxLength: 20
          example: "8542.31"
        tariff_type:
          type: string
          enum: [MFN, SECTION_301, SECTION_232, IEEPA]
          description: ประเภทภาษีที่ถูกยกเว้น
          example: IEEPA
        origin_country:
          type: string
          pattern: "^[A-Z]{2}$"
          default: CN
          example: CN
        effective_from:
          type: string
          format: date
          example: "2025-04-09"
        effective_to:
          type: string
          format: date
          nullable: true
          description: null = ยกเว้นถาวร
          example: null
        description:
          type: string
          maxLength: 500
          example: "Semiconductor ICs excluded from IEEPA"
        source_reference:
          type: string
          maxLength: 500
          example: "Executive Order 14257 Annex II"
    ExclusionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hts_code:
          type: string
        tariff_type:
          type: string
          enum: [MFN, SECTION_301, SECTION_232, IEEPA]
        origin_country:
          type: string
        effective_from:
          type: string
          format: date
        effective_to:
          type: string
          format: date
          nullable: true
        description:
          type: string
        source_reference:
          type: string
...